sched/fair: if sync flag ignored, try to place in same cluster
If the sync flag is ignored because the current cpu is not in
the affinity mask for the target of a sync wakeup (usually
binder call), prefer to place in the big cluster if
possible. The main case is a "top-app" task waking a
"foreground" task when the top-app task is running on
a CPU that is not in the foreground cpuset. This patch
causes the search order from big capacity cpu be used
when the sync flag failed.

Based on: https://github.com/aosp-mirror/kernel_msm/commit/580cecaa6bf324565694ddc688525ccd934d298f#diff-99e5389415cdd1d6dbf020da863c1ef5

Made into a sched feature

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>

[kailashrs: fixup for 4.9 and 845]
Signed-off-by: Kailash <kailash.sudhakar@gmail.com>