focaltech_touch: Optimize interrupt CPU usage
To process a touch interrupt, i2c reads are needed which make use of a
completion interrupt from the i2c controller (since i2c I/O is asynchronous).
This interrupt and the touch controller's interrupt are all that need to be
sped up. Additionally, speed up the i2c interrupt by marking it as
performance critical, which forces it to run on a fast CPU core.

Signed-off-by: Kailash <kailash.sudhakar@gmail.com>